+++ shared.go
@@ -6,13 +6,13 @@ func ms(ms int) time.Duration {
 	return time.Duration(ms) * time.Millisecond
 }
 
-func step(val int8) int8 {
+func step(val int) int {
 	// time.Sleep(ms(1))
 	return val + 1
 }
 
-func createArrayWithZeros(length int8) []int8 {
-	arr := make([]int8, length)
+func createArrayWithZeros(length int) []int {
+	arr := make([]int, length)
 
 	for i := range arr {
 		arr[i] = 0
